✨ Add RotarySensor class and remove Encoder class
